move epoch progress logic inside hooks
