+++ src/hooks/index.ts
@@ -8,6 +8,7 @@ export * from './useCheckVersion';
 export * from './useDebounce';
 export * from './useGetExplorerTitle';
 export * from './useGetHash';
+export * from './useGetEpochRemainingTime';
 export * from './useGetRemainingTime';
 export * from './useGetURLNetwork';
 export * from './useHasGrowthWidgets';
