+++ src/widgets/BlockProgressRing/BlockProgressRing.tsx
@@ -1,6 +1,6 @@
 import { ELLIPSIS } from 'appConstants';
 import { ProgressRing } from 'components';
-import { useFetchEpochProgress } from 'hooks/useFetchEpochProgress';
+import { useFetchEpochProgress } from 'hooks';
 import { WithClassnameType } from 'types';
 
 export const BlockProgressRing = ({ className }: WithClassnameType) => {
